Use new Target API (part 3) target.tab is the new target.value
Using 'value' didn't make much sense because it assumed that there was
only one thing that identified a target, which in the case of a remote
target with a host/port was untrue.
In addition if you've got the wrong target type, target.tab is fail-fast
where target.value is fail-confusing.