Updates to fit with review comments
The biggest change is to the construction API. I've left off the new/get
prefixes. (I started with 'new' to reflect "new Target", but it doesn't work
with plurals because it infers a single new thing, and this returns several
things, hence I switched to 'get', but several people didn't like that, hence
for/all) Also we're using TargetFactory rather than Target.