Fix some debugger tests: s/contentWindow/panelWin
Following the fix above in a merge from fx-team, I found the same thing in 5
other places, all of which were party to a commenting out. This fix was enough
for 3 of those files to be taken off the comment list.

Also I moved browser_dbg_createRemote.js up to be commented out in the normal
way.